+++ modules/book-class.js
@@ -1,0 +1,7 @@
+// book.js
+export class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
